renamed _utils.py, added constants for busyness and status
